+++ demos/CurlNoise/simulation/useGradientProgram/gradient.subtract.frag.js
@@ -1,0 +1,28 @@
+export default `
+  precision mediump float;
+  precision mediump sampler2D;
+
+  varying highp vec2 vUv;
+  varying highp vec2 vL;
+  varying highp vec2 vR;
+  varying highp vec2 vT;
+  varying highp vec2 vB;
+  uniform sampler2D uPressure;
+  uniform sampler2D uVelocity;
+
+  vec2 boundary (vec2 uv) {
+    return uv;
+    // uv = min(max(uv, 0.0), 1.0);
+    // return uv;
+  }
+
+  void main () {
+    float L = texture2D(uPressure, boundary(vL)).x;
+    float R = texture2D(uPressure, boundary(vR)).x;
+    float T = texture2D(uPressure, boundary(vT)).x;
+    float B = texture2D(uPressure, boundary(vB)).x;
+    vec2 velocity = texture2D(uVelocity, vUv).xy;
+    velocity.xy -= vec2(R - L, T - B);
+    gl_FragColor = vec4(velocity, 0.0, 1.0);
+  }
+`
